third_party/WebKit/LayoutTests/http/tests/webaudio/autoplay-crossorigin-expected.txt

CONSOLE WARNING: line 13: The AudioContext was not allowed to start. It must be resumed (or created) from a user gesture event handler.
CONSOLE WARNING: line 23: The AudioContext was not allowed to start. It must be resumed (or created) from a user gesture event handler.
CONSOLE WARNING: line 36: The AudioContext was not allowed to start. It must be resumed (or created) from a user gesture event handler.
CONSOLE ERROR: line 2684: Uncaught Error: assert_equals: stateAfterClick expected "running" but got "suspended"
This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught Error: assert_equals: stateAfterClick expected "running" but got "suspended"
PASS Verify that autoplaying Web Audio from a cross origin iframe is blocked by mediaPlaybackRequiresUserGesture
Harness: the test ran to completion.
